Add files via upload
A Data Quality Dashboard was built for Polen Capital as part of my Capstone Project. The brief was to build a dashboard for the business stewards so that they could a) get a snapshot of the data performance b) a trend history c) a tool to perform root-cause analysis. A Tachometer was built to give the overall data performance in a snapshot. Then a dynamic Timeline slicer was built to give a trend history by different categories. Finally, a Decomposition Tree was built to analyze the data failures. The dashboard is extremely dynamic in the sense, that it gives a bird's eye view and the functionality to investigate the root cause problem, at the level of the SQL query causing the problem. Please click this link below for a demo 
https://www.linkedin.com/safety/go?url=https%3A%2F%2Fwww.canva.com%2Fdesign%2FDAF6CdmdT-E%2Fyi9rky2Df0Z68FrWGQtxTg%2Fwatch%3Futm_content%3DDAF6CdmdT-E%26utm_campaign%3Dshare_your_design%26utm_medium%3Dlink%26utm_source%3Dshareyourdesignpanel&trk=flagship-messaging-web&messageThreadUrn=urn%3Ali%3AmessagingThread%3A2-ODFlZGZhNzctYmE4ZS00YWM2LTkwYTYtMGQ4YmMyM2Q5NWUxXzAxMg%3D%3D&lipi=urn%3Ali%3Apage%3Ad_flagship3_messaging_conversation_detail%3BvGUySkENTui7fLs8bKTIWA%3D%3D